chore #73 : 공유페이지 파일 정리
