+++ Outline/Outline/Source/Manager/PathGenerateManager.swift
@@ -6,6 +6,7 @@
 //
 
 import CoreLocation
+import MapKit
 import SwiftUI
 
 struct CanvasData {
@@ -90,5 +91,82 @@ final class PathGenerateManager {
             zeroX: minLon,
             zeroY: maxLat
             )
+    }
+}
+
+extension PathGenerateManager {
+    func caculateLinesInRect(
+        width: Double,
+        height: Double,
+        coordinates: [CLLocationCoordinate2D],
+        region: MKCoordinateRegion
+    ) -> some Shape {
+        let canvasData = calculateCanvaDataInRect(width: width, height: height, region: region)
+        var path = Path()
+        
+        let startPosition = calculateRelativePoint(coordinate: coordinates[0], canvasData: canvasData)
+        path.move(to: CGPoint(x: startPosition[0], y: -startPosition[1]))
+
+        for coordinate in coordinates {
+            let position = calculateRelativePoint(coordinate: coordinate, canvasData: canvasData)
+            path.addLine(to: CGPoint(x: position[0], y: -position[1]))
+        }
+        
+        return path
+    }
+    
+    private func calculateCanvaDataInRect(width: Double, height: Double, region: MKCoordinateRegion) -> CanvasDataForShare {
+        let minLon = region.center.longitude - region.span.longitudeDelta / 2
+        let maxLat = region.center.latitude + region.span.latitudeDelta / 2
+        
+        let calculatedHeight = region.span.latitudeDelta * 1000000
+        let calculatedWidth = region.span.longitudeDelta * 1000000
+        
+        let relativeWidthScale: Double = width / calculatedWidth
+        let relativeHeightScale: Double = height / calculatedHeight
+        
+        let fittedWidth = calculatedWidth * relativeWidthScale
+        let fittedHeight = calculatedHeight * relativeHeightScale
+        return CanvasDataForShare(
+            width: Int(fittedWidth),
+            height: Int(fittedHeight),
+            widthScale: relativeWidthScale,
+            heightScale: relativeHeightScale,
+            zeroX: minLon,
+            zeroY: maxLat
+            )
+    }
+    
+    func calculateDeltaAndCenter(coordinates: [CLLocationCoordinate2D]) -> MKCoordinateRegion {
+        var minLat: Double = 90
+        var maxLat: Double = -90
+        var minLon: Double = 180
+        var maxLon: Double = -180
+        for coordinate in coordinates {
+            if coordinate.latitude < minLat {
+                minLat = coordinate.latitude
+            }
+            if coordinate.latitude > maxLat {
+                maxLat = coordinate.latitude
+            }
+            if coordinate.longitude < minLon {
+                minLon = coordinate.longitude
+            }
+            if coordinate.longitude > maxLon {
+                maxLon = coordinate.longitude
+            }
+        }
+        let latitudeDelta = maxLat - minLat
+        let longitudeDelta = maxLon - minLon
+        let centerLatitude = (maxLat + minLat) / 2
+        let centerLongitude = (maxLon + minLon) / 2
+        
+        return MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: centerLatitude, longitude: centerLongitude), span: MKCoordinateSpan(latitudeDelta: latitudeDelta, longitudeDelta: longitudeDelta))
+    }
+    
+    private func calculateRelativePoint(coordinate: CLLocationCoordinate2D, canvasData: CanvasDataForShare) -> [Int] {
+        let posX = Int((coordinate.longitude - canvasData.zeroX) * canvasData.widthScale * 1000000)
+        let posY = Int((coordinate.latitude - canvasData.zeroY) * canvasData.heightScale * 1000000)
+        return [posX, posY]
     }
 }
